## Description
This PR resolves issue #132 by completing the dynamic-mask aware backward kernel refactor for Flash Dynamic Mask Attention.  
Key goals achieved:
1. Correctness: Eliminated stale shared-memory reuse races under aliased regions (sMask/sP and sBias/sdS) and removed scope-related undefined identifier errors (e.g., `dS`).
2. Dynamic Skip (Early Exit): Integrated block-level mask OR reduction to fully skip zero-score tiles (avoids 5 GEMMs + softmax path when a tile is fully masked).
3. Synchronization Slimming: Removed three redundant `__syncthreads()` while preserving required barriers around alias reuse and cooperative GEMMs.
4. Predicate Uniformity: Ensured `any_active` / `any_active_next` are CTA-uniform (`__syncthreads_or`) so that shared loads / stores are never partially executed.
5. Performance: Achieved 4.9×–6.2× speedups over PyTorch SDPA backward on RTX 4090 across window sizes (local attention window sweep).

## Type of Change
- [x] Performance optimization
- [x] CUDA kernel improvement
- [x] Bug fix (race + scope)
- [ ] New feature
- [ ] Breaking change
- [ ] Documentation update
- [ ] Code refactoring (partial – localized kernel restructuring)

## Related Issues
- Fixes #132

## Changes Made

### Code Changes
Backward CUDA kernel (`compute_dq_dk_dv_1colblock`):
- Added mask-first load + CTA-wide OR reduction to decide early skip.
- Replaced per-thread predicates with CTA-uniform `any_active` / `any_active_next`.
- Moved dS write path to reuse `acc_dp` buffer directly (removed cross-scope dependency).
- Removed 3 redundant barriers:
  * Prologue barrier before first `__syncthreads_or`
  * Post-`cp_async_wait<1>` barrier in V-in-regs path
  * Mask-next pre-OR barrier inside main loop
- Kept essential barriers before GEMM reading fully-populated tiles or before alias reuse of shared memory.
- Ensured safe alias transitions: (sdS ↔ sBias), (sP ↔ sMask), sdO reuse for dV GEMM.
- Converted fragmented dS row/col view to direct accumulator reinterpretation to lower register live range.
- Added conditional lazy loads (Bias only when `any_active`).

Other:
- Cleaned variable scope for `dS` to avoid compile error.
- Consistent cp.async sequence: copy → fence → wait → (optional) barrier → OR → conditional dependent copies.

### Documentation
- (Pending) Will add a short design note summarizing backward early-skip math & synchronization invariants.
- (Optional) Update README performance matrix (not part of this PR yet).

### Testing
No Python API surface change. No dependency changes.

## Testing
Performed:
- Functional equivalence vs PyTorch SDPA backward across:
  * Multiple window sizes (32 → full length)
  * Causal and non-causal (causal path sampled)
  * Head dimension D=32 (baseline tuned path)
- Randomized sparsity patterns (window-based masking)
- Repeated runs (warmup + timed) to stabilize measurement
- Checked for NaN/Inf emergence (none observed)
- Verified no race by running with `CUDA_LAUNCH_BLOCKING=1` and cuda-memcheck (spot tests)

Planned / Deferred:
- Multi-architecture validation (SM80, SM90)
- Larger head dims (64, 128)
- Gradient equivalence harness extension to atomic-accum (Seq_parallel path)

### Test Configuration
- OS: (Container) Ubuntu base (inside docker)
- Python: 3.12.x (inferred from `.cpython-312` build)
- PyTorch: 2.8.0a0+5228986c39.nv25.05
- CUDA: Driver compatible with RTX 4090 (SM 89)
- GPU: NVIDIA GeForce RTX 4090
- Seed: 42
- Runs: Warmup 2, Timed 3

## Performance Impact

### Benchmark Protocol
Command:
```
python benchmarks/backward_performance.py --test-type sdpa-vs-cuda
```
Mode: Local window attention sweeping window size (W) while keeping Q=K=16384, B=1, Hq=2, Hkv=1, D=32, causal (C).

### Results (Averaged over 3 runs)

| Config (B1 Hq2 Hkv1 Q16384 K16384 D32 C) | Window (W) | SDPA-BWD (ms) | CUDA-BWD (ms) | Speedup |
|------------------------------------------|------------|---------------|---------------|---------|
|                                          | 32         | 28.42         | 4.64          | 6.1×    |
|                                          | 64         | 28.27         | 4.57          | 6.2×    |
|                                          | 128        | 28.36         | 4.67          | 6.1×    |
|                                          | 256        | 28.52         | 4.69          | 6.1×    |
|                                          | 512        | 28.67         | 5.28          | 5.4×    |
|                                          | 1024       | 28.47         | 4.83          | 5.9×    |
|                                          | 2048       | 28.56         | 4.86          | 5.9×    |
|                                          | 4096       | 28.39         | 4.91          | 5.8×    |
|                                          | 8192       | 28.37         | 5.68          | 5.0×    |
|                                          | 16384 (full) | 28.50       | 5.87          | 4.9×    |

Observations:
- SDPA baseline remains nearly flat (compute-bound softmax + matmuls).
- Our backward stays ~4.6–5.0 ms for small/mid windows; degradation for large windows is expected as mask density approaches dense case and skip opportunities vanish.
- Speedup highest at small to medium window (sparser effective attention region).
- Removing redundant barriers shaved ~2–4% from initial tuned version (micro-profiling—numbers not shown here, can add if needed).

### Before (Legacy)
No earlier native dynamic mask backward existed in-tree (baseline = PyTorch SDPA). If you want a “Before” column for internal review later, we can add raw times from pre-synchronization-slimming commit.

## Breaking Changes
None.
- API unchanged.
- Memory footprint unchanged (no extra persistent buffers).
- Determinism unaffected (barrier removals do not reorder floating ops across independent tiles).

## Checklist
- [x] Code follows style guidelines
- [x] Self-reviewed
- [x] Critical sections commented (synchronization + alias rationale)
- [x] No new warnings in CUDA compilation (spot check)
- [x] Existing benchmarks pass
- [x] No perf regression (improvement verified)
- [ ] Added doc note (will follow-up)
- [ ] Multi-arch test (deferred)
- [ ] Larger head dims validation (deferred)

### CUDA-specific
- [x] Kernels compile without warnings (baseline environment)
- [x] Tested on SM 89 (RTX 4090)
- [x] Shared-memory aliasing guarded by barriers
- [x] Removed only provably redundant barriers
- [ ] Memory usage profiling (deferred; expected neutral)
- [ ] SM80 / SM90 validation (planned)

## Additional Notes
Future work (tracked or to be filed):
1. Double-buffering pipeline for backward (overlap next mask + dO load with current GEMMs).
2. Bit-packed mask to reduce L2 bandwidth at mid sparsity.
3. Instrumentation counters (tiles_skipped / tiles_total) for adaptive heuristics.
4. Multi-head / high-D tuning (ensure register pressure stable when D=64/128).
5. Extend Triton/Flex backward parity benchmarks once those variants support dynamic mask skip semantics.

Safety rationale for removed barriers:
- Every removed `__syncthreads()` was immediately followed by a `__syncthreads_or`, or only guarded accesses to per-thread exclusive SMEM slices post `cp_async_wait`.
- All GEMM consumption points still have an immediately preceding full CTA barrier ensuring tile completeness.